audit not working with double

+++ src/Client.h
@@ -11,7 +11,7 @@
 
 class Client
 {
-    typedef Givaro::ModularBalanced<int64_t> Field;
+    typedef Givaro::ModularBalanced<double> Field;
 private:
 
     Field F;
